requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
qdrant-client==1.12.1
sentence-transformers==3.3.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.3
email-validator==2.1.0